LinkedList generic completed

+++ src/com/bridgeLabz/basicCorePrograms/PowerOfTwo.java
@@ -11,12 +11,12 @@ public class PowerOfTwo {
 		// 11
 		System.out.println("Enter the power of Two");
 		int power = sc.nextInt();
-		int result = Util.powerOfTwo(power);
+		double result = Util.powerOfTwo(power);
 
 		if (power >= 1 && power <= 31) {
 			System.out.println("2 ^ " + power + " = " + result);
 
-			if (Util.isLeapYear(result) == true) {
+			if (Util.isLeapYear((int)result) == true) {
 				System.out.println(result + " is a leap year");
 			} else {
 				System.out.println(result + " not a leap year");
